feature oriented architecture for User domain
