--- config/routes/login.yaml
@@ -1,9 +1,0 @@
-app_login:
-    path: /
-    controller: App\Controller\Login\Security::handle
-    methods: [GET, POST]
-
-login_success:
-    path: /success
-    controller: App\Controller\Login\SignInSuccessful::handle
-    methods: [GET]
