--- src/Exceptions/EmptyPlainPasswordException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exceptions;
-
-class EmptyPlainPasswordException extends \Exception
-{
-}
