--- src/Exceptions/User/UnableToCreateUserException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exceptions\User;
-
-class UnableToCreateUserException extends \LogicException
-{
-}
