--- src/Exceptions/User/UnableToDeleteUserException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exceptions\User;
-
-class UnableToDeleteUserException extends \LogicException
-{
-}
