--- src/Exceptions/User/UnableToSaveUserException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exceptions\User;
-
-class UnableToSaveUserException extends \LogicException
-{
-}
