--- src/Exceptions/ValueObject/Username/InvalidUsernameException.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exceptions\ValueObject\Username;
-
-use Throwable;
-
-class InvalidUsernameException extends \LogicException
-{
-    public function __construct(string $message = '', int $code = 0, Throwable $previous = null)
-    {
-        parent::__construct($message, $code, $previous);
-    }
-}
