--- src/Form/LoginType.php
@@ -1,25 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Form;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\Extension\Core\Type\EmailType;
-use Symfony\Component\Form\Extension\Core\Type\PasswordType;
-
-final class LoginType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $formBuilder, array $options): void
-    {
-        $formBuilder
-            ->add('email', EmailType::class, [
-                'required' => true,
-            ])
-            ->add('password', PasswordType::class, [
-                'required' => true,
-            ])
-        ;
-    }
-}
