+++ src/User/Create/UnableToCreateUserException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\User\Create;
+
+class UnableToCreateUserException extends \LogicException
+{
+}
