+++ src/User/Delete/UnableToDeleteUserException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\User\Delete;
+
+class UnableToDeleteUserException extends \LogicException
+{
+}
