+++ src/User/Login/EmptyPlainPasswordException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\User\Login;
+
+class EmptyPlainPasswordException extends \Exception
+{
+}
