+++ src/User/Login/LoginFormType.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\User\Login;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+
+final class LoginFormType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $formBuilder, array $options): void
+    {
+        $formBuilder
+            ->add('email', EmailType::class, [
+                'required' => true,
+            ])
+            ->add('password', PasswordType::class, [
+                'required' => true,
+            ])
+        ;
+    }
+}
