+++ src/User/Resources/config/doctrine/User.orm.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="App\User\User" table="app_user" repositoryClass="App\User\Doctrine\UserRepository">
+        <id name="id" type="guid">
+            <generator strategy="NONE"/>
+        </id>
+        <field name="email" length="320" unique="true"/>
+        <field name="username" length="50"/>
+        <field name="roles" type="json"/>
+        <field name="password" length="320"/>
+    </entity>
+</doctrine-mapping>
