+++ src/User/Update/UnableToSaveUserException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\User\Update;
+
+class UnableToSaveUserException extends \LogicException
+{
+}
